backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.3
pydantic-settings==2.10.1
python-multipart==0.0.20
httpx==0.28.1
openai==1.108.1
groq==0.31.0
langfuse==3.4.0
openlit==1.34.0
PyJWT==2.10.1
SQLAlchemy==2.0.41
PyMySQL==1.1.1
bcrypt==5.0.0
cryptography==44.0.1
